Separate transport related features from Swarm<T>
[skip changelog]
