+++ Libplanet/Net/Swarm.cs
@@ -6,7 +6,6 @@ using System.Collections.Immutable;
 using System.IO;
 using System.Linq;
 using System.Net;
-using System.Net.Sockets;
 using System.Security.Cryptography;
 using System.Threading;
 using System.Threading.Tasks;
@@ -19,27 +18,16 @@ using Libplanet.Crypto;
 using Libplanet.Net.Messages;
 using Libplanet.Net.Protocols;
 using Libplanet.Store;
-using Libplanet.Stun;
 using Libplanet.Tx;
-using NetMQ;
-using NetMQ.Sockets;
 using Nito.AsyncEx;
 using Serilog;
 using Serilog.Events;
 
 namespace Libplanet.Net
 {
-    public class Swarm<T> : ISwarm, IDisposable
+    public class Swarm<T> : IDisposable
         where T : IAction, new()
     {
-        private static readonly TimeSpan TurnAllocationLifetime =
-            TimeSpan.FromSeconds(777);
-
-        // TURN Permission lifetime was defined in RFC 5766
-        // see also https://tools.ietf.org/html/rfc5766#section-8
-        private static readonly TimeSpan TurnPermissionLifetime =
-            TimeSpan.FromMinutes(5);
-
         private static readonly TimeSpan MaxTimeout = TimeSpan.FromSeconds(150);
         private static readonly TimeSpan BlockRecvTimeout = TimeSpan.FromSeconds(15);
         private static readonly TimeSpan TxRecvTimeout = TimeSpan.FromSeconds(3);
@@ -47,30 +35,16 @@ namespace Libplanet.Net
         private readonly PrivateKey _privateKey;
         private readonly int _appProtocolVersion;
 
-        private readonly AsyncLock _runningMutex;
         private readonly AsyncLock _blockSyncMutex;
-        private readonly string _host;
-        private readonly IList<IceServer> _iceServers;
+        private readonly AsyncLock _runningMutex;
 
         private readonly ILogger _logger;
 
-        private RouterSocket _router;
-        private NetMQQueue<Message> _replyQueue;
-        private NetMQQueue<(Address?, Message)> _broadcastQueue;
-        private AsyncCollection<MessageRequest> _requests;
-        private NetMQPoller _poller;
+        private readonly ITransport _transport;
+        private readonly IStore _store;
 
-        private TaskCompletionSource<object> _runningEvent;
-        private int? _listenPort;
-        private TurnClient _turnClient;
-        private bool _behindNAT;
         private CancellationTokenSource _workerCancellationTokenSource;
-        private CancellationTokenSource _runtimeCancellationTokenSource;
         private CancellationToken _cancellationToken;
-        private IPAddress _publicIPAddress;
-        private Task _runtimeProcessor;
-
-        private IStore _store;
 
         static Swarm()
         {
@@ -119,8 +93,6 @@ namespace Libplanet.Net
             IEnumerable<IceServer> iceServers = null,
             EventHandler<DifferentProtocolVersionEventArgs> differentVersionPeerEncountered = null)
         {
-            Running = false;
-
             BlockChain = blockChain ?? throw new ArgumentNullException(nameof(blockChain));
             _store = BlockChain.Store;
             _privateKey = privateKey ?? throw new ArgumentNullException(nameof(privateKey));
@@ -133,116 +105,48 @@ namespace Libplanet.Net
             TxReceived = new AsyncAutoResetEvent();
             BlockAppended = new AsyncAutoResetEvent();
             BlockReceived = new AsyncAutoResetEvent();
-            DifferentVersionPeerEncountered = differentVersionPeerEncountered;
 
             _blockSyncMutex = new AsyncLock();
             _runningMutex = new AsyncLock();
 
-            _host = host;
-            _listenPort = listenPort;
             _appProtocolVersion = appProtocolVersion;
-
-            if (_host != null && _listenPort is int listenPortAsInt)
-            {
-                EndPoint = new DnsEndPoint(_host, listenPortAsInt);
-            }
-
-            _iceServers = iceServers?.ToList();
-            if (_host == null && (_iceServers == null || !_iceServers.Any()))
-            {
-                throw new ArgumentException(
-                    $"Swarm requires either {nameof(host)} or " +
-                    $"{nameof(iceServers)}."
-                );
-            }
 
             string loggerId = _privateKey.PublicKey.ToAddress().ToHex();
             _logger = Log.ForContext<Swarm<T>>()
                 .ForContext("SwarmId", loggerId);
 
-            Protocol = new KademliaProtocol(
-                this,
-                _privateKey.PublicKey.ToAddress(),
+            _transport = new NetMQTransport(
+                _privateKey,
                 _appProtocolVersion,
-                _logger,
                 tableSize,
-                bucketSize);
-
-            _requests = new AsyncCollection<MessageRequest>();
-            _runtimeCancellationTokenSource = new CancellationTokenSource();
-            _runtimeProcessor = Task.Factory.StartNew(
-                () =>
-                {
-                    // Ignore NetMQ related exceptions during NetMQRuntime.Dispose() to stabilize
-                    // tests
-                    try
-                    {
-                        using (var runtime = new NetMQRuntime())
-                        {
-                            var workerTasks = new Task[workers];
-
-                            for (int i = 0; i < workers; i++)
-                            {
-                                workerTasks[i] = ProcessRuntime(
-                                    _runtimeCancellationTokenSource.Token
-                                );
-                            }
-
-                            runtime.Run(workerTasks);
-                        }
-                    }
-                    catch (NetMQException e)
-                    {
-                        _logger.Error(
-                            e,
-                            $"NetMQException occurred in {nameof(_runtimeProcessor)}."
-                        );
-                    }
-                    catch (ObjectDisposedException e)
-                    {
-                        _logger.Error(
-                            e,
-                            $"ObjectDisposedException occurred in {nameof(_runtimeProcessor)}."
-                        );
-                    }
-                },
-                CancellationToken.None,
-                TaskCreationOptions.DenyChildAttach | TaskCreationOptions.LongRunning,
-                TaskScheduler.Default
-            );
+                bucketSize,
+                workers,
+                host,
+                listenPort,
+                iceServers,
+                differentVersionPeerEncountered,
+                ProcessMessageHandler,
+                _logger);
         }
 
         ~Swarm()
         {
             // FIXME If possible, we should stop Swarm appropriately here.
-            if (Running)
+            if (_transport is NetMQTransport p && p.Running)
             {
                 _logger.Warning(
-                    "Swarm is scheduled to destruct, but it's still running.");
+                    "Swarm is scheduled to destruct, but NetMQTransport progress is still running."
+                );
             }
         }
 
-        /// <summary>
-        /// The <see cref="EventHandler" /> triggered when the different version of
-        /// <see cref="Peer" /> is discovered.
-        /// </summary>
-        public event EventHandler<DifferentProtocolVersionEventArgs>
-            DifferentVersionPeerEncountered;
+        public bool Running => _transport is NetMQTransport p && p.Running;
 
-        public DnsEndPoint EndPoint { get; private set; }
+        public DnsEndPoint EndPoint => (_transport as NetMQTransport)?.EndPoint;
 
         public Address Address => _privateKey.PublicKey.ToAddress();
 
-        public Peer AsPeer => EndPoint is null
-            ? new Peer(
-                _privateKey.PublicKey,
-                _appProtocolVersion,
-                _publicIPAddress)
-            : new BoundPeer(
-                _privateKey.PublicKey,
-                EndPoint,
-                _appProtocolVersion,
-                _publicIPAddress);
+        public Peer AsPeer => (_transport as NetMQTransport)?.AsPeer;
 
         public DateTimeOffset LastReceived { get; private set; }
 
@@ -252,33 +156,15 @@ namespace Libplanet.Net
             private set;
         }
 
+        public IEnumerable<BoundPeer> Peers => _transport.Peers();
+
         /// <summary>
         /// The <see cref="BlockChain{T}"/> instance this <see cref="Swarm{T}"/> instance
         /// synchronizes with.
         /// </summary>
         public BlockChain<T> BlockChain { get; private set; }
 
-        /// <summary>
-        /// Whether this <see cref="Swarm{T}"/> instance is running.
-        /// </summary>
-        public bool Running
-        {
-            get => _runningEvent.Task.Status == TaskStatus.RanToCompletion;
-
-            private set
-            {
-                if (value)
-                {
-                    _runningEvent.TrySetResult(null);
-                }
-                else
-                {
-                    _runningEvent = new TaskCompletionSource<object>();
-                }
-            }
-        }
-
-        public IEnumerable<BoundPeer> Peers => Protocol.Peers;
+        internal IProtocol Protocol => (_transport as NetMQTransport)?.Protocol;
 
         internal AsyncAutoResetEvent TxReceived { get; }
 
@@ -288,8 +174,6 @@ namespace Libplanet.Net
         internal AsyncAutoResetEvent BlockAppended { get; }
 
         internal TimeSpan BlockHashRecvTimeout { get; set; } = TimeSpan.FromSeconds(30);
-
-        internal IProtocol Protocol { get; private set; }
 
         // FIXME: We need some sort of configuration method for it.
         internal int FindNextHashesChunkSize { get; set; } = 500;
@@ -305,15 +189,13 @@ namespace Libplanet.Net
         /// <summary>
         /// Waits until this <see cref="Swarm{T}"/> instance gets started to run.
         /// </summary>
-        /// <returns>A <see cref="Task"/> completed when <see cref="Running"/>
+        /// <returns>A <see cref="Task"/> completed when <see cref="NetMQTransport.Running"/>
         /// property becomes <c>true</c>.</returns>
-        public Task WaitForRunningAsync() => _runningEvent.Task;
+        public Task WaitForRunningAsync() => (_transport as NetMQTransport)?.WaitForRunningAsync();
 
         public void Dispose()
         {
-            _runtimeCancellationTokenSource.Cancel();
-
-            _runtimeProcessor.Wait();
+            (_transport as NetMQTransport)?.Dispose();
         }
 
         public async Task StopAsync(
@@ -332,26 +214,7 @@ namespace Libplanet.Net
             _logger.Debug("Stopping...");
             using (await _runningMutex.LockAsync())
             {
-                if (Running)
-                {
-                    await Task.Delay(waitFor, cancellationToken);
-
-                    _broadcastQueue.ReceiveReady -= DoBroadcast;
-                    _replyQueue.ReceiveReady -= DoReply;
-                    _router.ReceiveReady -= ReceiveMessage;
-
-                    if (_poller.IsRunning)
-                    {
-                        _poller.Dispose();
-                    }
-
-                    _broadcastQueue.Dispose();
-                    _replyQueue.Dispose();
-                    _router.Dispose();
-                    _turnClient?.Dispose();
-
-                    Running = false;
-                }
+                await _transport.StopAsync(waitFor, cancellationToken);
             }
 
             _logger.Debug("Stopped.");
@@ -407,78 +270,12 @@ namespace Libplanet.Net
             EventHandler<PreloadBlockDownloadFailEventArgs> preloadBlockDownloadFailed = null,
             CancellationToken cancellationToken = default(CancellationToken))
         {
-            if (Running)
-            {
-                throw new SwarmException("Swarm is already running.");
-            }
-
-            _router = new RouterSocket();
-            _router.Options.RouterHandover = true;
-
-            if (_host is null && !(_iceServers is null))
-            {
-                _turnClient = await IceServer.CreateTurnClient(_iceServers);
-            }
-
-            if (_listenPort == null)
-            {
-                _listenPort = _router.BindRandomPort("tcp://*");
-            }
-            else
-            {
-                _router.Bind($"tcp://*:{_listenPort}");
-            }
-
-            _logger.Information($"Listen on {_listenPort}");
-
+            List<Task> tasks = new List<Task>();
             _workerCancellationTokenSource = new CancellationTokenSource();
-            CancellationToken workerCancellationToken =
-                CancellationTokenSource.CreateLinkedTokenSource(
+            _cancellationToken = CancellationTokenSource.CreateLinkedTokenSource(
                     _workerCancellationTokenSource.Token, cancellationToken
                 ).Token;
-            _cancellationToken = workerCancellationToken;
-            var tasks = new List<Task>();
-
-            if (!(_turnClient is null))
-            {
-                _publicIPAddress = (await _turnClient.GetMappedAddressAsync()).Address;
-
-                if (await _turnClient.IsBehindNAT())
-                {
-                    _behindNAT = true;
-                }
-            }
-
-            if (_behindNAT)
-            {
-                IPEndPoint turnEp = await _turnClient.AllocateRequestAsync(
-                    TurnAllocationLifetime
-                );
-                EndPoint = new DnsEndPoint(turnEp.Address.ToString(), turnEp.Port);
-
-                // FIXME should be parameterized
-                tasks.Add(BindingProxies(_cancellationToken));
-                tasks.Add(BindingProxies(_cancellationToken));
-                tasks.Add(BindingProxies(_cancellationToken));
-                tasks.Add(RefreshAllocate(_cancellationToken));
-                tasks.Add(RefreshPermissions(_cancellationToken));
-            }
-            else if (_host is null)
-            {
-                EndPoint = new DnsEndPoint(_publicIPAddress.ToString(), _listenPort.Value);
-            }
-            else
-            {
-                EndPoint = new DnsEndPoint(_host, _listenPort.Value);
-            }
-
-            _replyQueue = new NetMQQueue<Message>();
-            _broadcastQueue = new NetMQQueue<(Address?, Message)>();
-            _poller = new NetMQPoller { _router, _replyQueue, _broadcastQueue };
-
-            _router.ReceiveReady += ReceiveMessage;
-            _replyQueue.ReceiveReady += DoReply;
-            _broadcastQueue.ReceiveReady += DoBroadcast;
+            await _transport.StartAsync(_cancellationToken);
 
             _logger.Debug("Starting swarm...");
             _logger.Debug("Peer information : {Peer}", AsPeer);
@@ -490,36 +287,12 @@ namespace Libplanet.Net
                     render: true,
                     cancellationToken: _cancellationToken
                 );
-                Running = true;
             }
 
             try
             {
+                tasks.Add(_transport.RunAsync(_cancellationToken));
                 tasks.Add(BroadcastTxAsync(broadcastTxInterval, _cancellationToken));
-                tasks.Add(
-                    RefreshTableAsync(
-                        TimeSpan.FromSeconds(10),
-                        TimeSpan.FromSeconds(10),
-                        _cancellationToken));
-                tasks.Add(RebuildConnectionAsync(TimeSpan.FromMinutes(30), _cancellationToken));
-                tasks.Add(
-                    Task.Run(() =>
-                    {
-                        // Ignore NetMQ related exceptions during NetMQPoller.Run() to stabilize
-                        // tests.
-                        try
-                        {
-                            _poller.Run();
-                        }
-                        catch (TerminatingException)
-                        {
-                            _logger.Error($"TerminatingException occurred in {nameof(_poller)}");
-                        }
-                        catch (ObjectDisposedException)
-                        {
-                            _logger.Error($"ObjectDisposedException occurred in {nameof(_poller)}");
-                        }
-                    }));
                 _logger.Debug("Swarm started.");
 
                 await await Task.WhenAny(tasks);
@@ -581,7 +354,7 @@ namespace Libplanet.Net
 
             IEnumerable<BoundPeer> peers = seedPeers.OfType<BoundPeer>();
 
-            await Protocol.BootstrapAsync(
+            await _transport.BootstrapAsync(
                 peers.ToImmutableList(),
                 pingSeedTimeout,
                 findNeighborsTimeout,
@@ -601,26 +374,7 @@ namespace Libplanet.Net
 
         public string TraceTable()
         {
-            return Protocol is null ? string.Empty : Protocol.Trace();
-        }
-
-        async Task<Message> ISwarm.SendMessageWithReplyAsync(
-            BoundPeer peer,
-            Message message,
-            TimeSpan? timeout,
-            CancellationToken cancellationToken)
-        {
-            IEnumerable<Message> replies =
-                await SendMessageWithReplyAsync(peer, message, timeout, 1, cancellationToken);
-            Message reply = replies.First();
-            ValidateSender(reply.Remote);
-
-            return reply;
-        }
-
-        void ISwarm.ReplyMessage(Message message)
-        {
-            _replyQueue.Enqueue(message);
+            return _transport is null ? string.Empty : (_transport as NetMQTransport)?.Trace();
         }
 
 #pragma warning disable MEN002 // Line is too long
@@ -687,13 +441,12 @@ namespace Libplanet.Net
             TimeSpan? timeout,
             CancellationToken cancellationToken)
         {
-            KademliaProtocol kp = (KademliaProtocol)Protocol;
-            return await kp.FindSpecificPeerAsync(
-                new ConcurrentBag<BoundPeer>(),
+            NetMQTransport netMQTransport = (NetMQTransport)_transport;
+            return await netMQTransport.FindSpecificPeerAsync(
                 target,
-                viaPeer,
-                depth,
                 searchAddress,
+                depth,
+                viaPeer,
                 timeout,
                 cancellationToken);
         }
@@ -885,9 +638,9 @@ namespace Libplanet.Net
             TimeSpan? timeout,
             CancellationToken cancellationToken = default(CancellationToken))
         {
-            if (Protocol is null)
+            if (_transport is null)
             {
-                throw new ArgumentNullException(nameof(Protocol));
+                throw new ArgumentNullException(nameof(_transport));
             }
 
             if (cancellationToken == default(CancellationToken))
@@ -895,134 +648,7 @@ namespace Libplanet.Net
                 cancellationToken = _cancellationToken;
             }
 
-            try
-            {
-                KademliaProtocol kp = (KademliaProtocol)Protocol;
-
-                var tasks = new List<Task>();
-                foreach (Peer peer in peers)
-                {
-                    if (peer is BoundPeer boundPeer)
-                    {
-                        tasks.Add(kp.PingAsync(
-                            boundPeer,
-                            timeout: timeout,
-                            cancellationToken: cancellationToken));
-                    }
-                }
-
-                _logger.Verbose("Trying to ping all {PeersNumber} peers.", tasks.Count);
-                await Task.WhenAll(tasks);
-                _logger.Verbose("Update complete.");
-            }
-            catch (DifferentAppProtocolVersionException)
-            {
-                _logger.Debug("Different version encountered during AddPeersAsync().");
-            }
-            catch (TimeoutException)
-            {
-                _logger.Debug(
-                    $"Timeout occurred during {nameof(AddPeersAsync)}() after {timeout}.");
-                throw;
-            }
-            catch (TaskCanceledException)
-            {
-                _logger.Debug($"Task is cancelled during {nameof(AddPeersAsync)}().");
-            }
-            catch (Exception e)
-            {
-                _logger.Error(
-                    e,
-                    $"Unexpected exception occurred during {nameof(AddPeersAsync)}().");
-                throw;
-            }
-        }
-
-        internal async Task SendMessageAsync(BoundPeer peer, Message message)
-        {
-            await SendMessageWithReplyAsync(peer, message, TimeSpan.FromSeconds(3), 0);
-        }
-
-        internal virtual async Task<IEnumerable<Message>> SendMessageWithReplyAsync(
-            BoundPeer peer,
-            Message message,
-            TimeSpan? timeout,
-            int expectedResponses,
-            CancellationToken cancellationToken = default(CancellationToken)
-        )
-        {
-            if (_behindNAT)
-            {
-                await CreatePermission(peer);
-            }
-
-            Guid reqId = Guid.NewGuid();
-            try
-            {
-                _logger.Verbose(
-                    "Enqueue a request {RequestId} to {PeerAddress}: {Message}.",
-                    reqId,
-                    peer.Address,
-                    message
-                );
-                var tcs = new TaskCompletionSource<IEnumerable<Message>>();
-                await _requests.AddAsync(
-                    new MessageRequest(reqId, message, peer, timeout, expectedResponses, tcs)
-                );
-                _logger.Verbose(
-                    "Enqueued a request {RequestId} to {PeerAddress}: {Message}.",
-                    reqId,
-                    peer.Address,
-                    message
-                );
-
-                var reply = (await tcs.Task).ToList();
-                const string logMsg =
-                    "Received {ReplyMessageCount} reply messages to {RequestId} " +
-                    "from {PeerAddress}: {ReplyMessages}.";
-                _logger.Debug(logMsg, reply.Count, reqId, peer.Address, reply);
-
-                return reply;
-            }
-            catch (DifferentAppProtocolVersionException e)
-            {
-                const string logMsg =
-                    "{PeerAddress} sent a reply to {RequestId} with " +
-                    "a different app protocol version; " +
-                    "expected: {ExpectedVersion}; actual: {ActualVersion}.";
-                _logger.Error(e, logMsg, peer.Address, reqId, e.ExpectedVersion, e.ActualVersion);
-                throw;
-            }
-            catch (TimeoutException)
-            {
-                _logger.Debug(
-                    $"{nameof(Swarm<T>)}.{nameof(SendMessageWithReplyAsync)}() timed out " +
-                    "after {Timeout} of waiting a reply to {RequestId} from {PeerAddress}.",
-                    timeout,
-                    reqId,
-                    peer.Address
-                );
-                throw;
-            }
-            catch (TaskCanceledException)
-            {
-                _logger.Debug(
-                    $"{nameof(Swarm<T>)}.{nameof(SendMessageWithReplyAsync)}() was cancelled to " +
-                    "wait a reply to {RequestId} from {PeerAddress}.",
-                    reqId,
-                    peer.Address
-                );
-                throw;
-            }
-            catch (Exception e)
-            {
-                var msg =
-                    $"{nameof(Swarm<T>)}.{nameof(SendMessageWithReplyAsync)}() encountered " +
-                    "an unexpected exception during sending a request {RequestId} to " +
-                    "{PeerAddress} and waiting a reply to it: {Exception}.";
-                _logger.Error(e, msg, reqId, peer.Address, e);
-                throw;
-            }
+            await ((NetMQTransport)_transport).AddPeersAsync(peers, timeout, cancellationToken);
         }
 
         internal async Task<IEnumerable<HashDigest<SHA256>>>
@@ -1035,15 +661,12 @@ namespace Libplanet.Net
         {
             var request = new GetBlockHashes(locator, stop);
 
-            Message parsedMessage = await (this as ISwarm).SendMessageWithReplyAsync(
+            Message parsedMessage = await _transport.SendMessageWithReplyAsync(
                 peer,
                 request,
                 timeout: BlockHashRecvTimeout,
                 cancellationToken: token
             );
-
-            Protocol.ReceiveMessage(parsedMessage);
-            ValidateSender(parsedMessage.Remote);
 
             if (parsedMessage is BlockHashes blockHashes)
             {
@@ -1079,7 +702,7 @@ namespace Libplanet.Net
                     blockRecvTimeout = MaxTimeout;
                 }
 
-                IEnumerable<Message> replies = await SendMessageWithReplyAsync(
+                IEnumerable<Message> replies = await _transport.SendMessageWithReplyAsync(
                     peer,
                     request,
                     blockRecvTimeout,
@@ -1089,8 +712,6 @@ namespace Libplanet.Net
 
                 foreach (Message message in replies)
                 {
-                    Protocol.ReceiveMessage(message);
-                    ValidateSender(message.Remote);
                     if (message is Messages.Blocks blockMessage)
                     {
                         IList<byte[]> payloads = blockMessage.Payloads;
@@ -1133,7 +754,7 @@ namespace Libplanet.Net
                     txRecvTimeout = MaxTimeout;
                 }
 
-                IEnumerable<Message> replies = await SendMessageWithReplyAsync(
+                IEnumerable<Message> replies = await _transport.SendMessageWithReplyAsync(
                     peer,
                     request,
                     txRecvTimeout,
@@ -1143,8 +764,6 @@ namespace Libplanet.Net
 
                 foreach (Message message in replies)
                 {
-                    Protocol.ReceiveMessage(message);
-                    ValidateSender(message.Remote);
                     if (message is Messages.Tx parsed)
                     {
                         Transaction<T> tx = Transaction<T>.FromBencodex(
@@ -1180,36 +799,7 @@ namespace Libplanet.Net
 
         private void BroadcastMessage(Address? except, Message message)
         {
-            _broadcastQueue.Enqueue((except, message));
-        }
-
-        private async Task BindingProxies(CancellationToken cancellationToken)
-        {
-            while (!cancellationToken.IsCancellationRequested)
-            {
-                try
-                {
-#pragma warning disable IDE0067  // We'll dispose of `stream` in proxy task.
-                    NetworkStream stream = await _turnClient.AcceptRelayedStreamAsync();
-#pragma warning restore IDE0067
-
-                    // TODO We should expose the interface so that library users
-                    // can limit / manage the task.
-#pragma warning disable CS4014
-                    Task.Run(async () =>
-                    {
-                        using (var proxy = new NetworkStreamProxy(stream))
-                        {
-                            await proxy.StartAsync(IPAddress.Loopback, _listenPort.Value);
-                        }
-                    }).ContinueWith(_ => stream.Dispose());
-#pragma warning restore CS4014
-                }
-                catch (Exception e)
-                {
-                    _logger.Error(e, "An unexpected exception occurred. try again...");
-                }
-            }
+            _transport.BroadcastMessage(except, message);
         }
 
         private System.Collections.Async.IAsyncEnumerable<(BoundPeer, Pong)> DialToExistingPeers(
@@ -1222,7 +812,7 @@ namespace Libplanet.Net
                 {
                     try
                     {
-                        Message reply = await (this as ISwarm).SendMessageWithReplyAsync(
+                        Message reply = await _transport.SendMessageWithReplyAsync(
                             peer,
                             new Ping(),
                             dialTimeout,
@@ -1230,7 +820,6 @@ namespace Libplanet.Net
                         );
                         if (reply is Pong pong)
                         {
-                            Protocol.ReceiveMessage(reply);
                             await yield.ReturnAsync((peer, pong));
                         }
                     }
@@ -1313,15 +902,13 @@ namespace Libplanet.Net
                     Message reply;
                     try
                     {
-                        reply = await (this as ISwarm).SendMessageWithReplyAsync(
+                        reply = await _transport.SendMessageWithReplyAsync(
                             peer,
                             request,
                             timeout: RecentStateRecvTimeout,
                             cancellationToken: cancellationToken
                         );
 
-                        // Do not call Protocol.ReceiveMessage since swarm is not started
-                        // when synchronizing previous states.
                         _logger.Debug("Received recent states from a peer ({Peer}).", peer);
                     }
                     catch (TimeoutException e)
@@ -1332,8 +919,6 @@ namespace Libplanet.Net
                         );
                         break;
                     }
-
-                    ValidateSender(reply.Remote);
 
                     if (reply is RecentStates recentStates && !recentStates.Missing)
                     {
@@ -1472,58 +1057,6 @@ namespace Libplanet.Net
             _logger.Debug("Finished to execute actions.");
         }
 
-        private async Task RefreshAllocate(CancellationToken cancellationToken)
-        {
-            TimeSpan lifetime = TurnAllocationLifetime;
-            while (!cancellationToken.IsCancellationRequested)
-            {
-                try
-                {
-                    await Task.Delay(lifetime - TimeSpan.FromMinutes(1), cancellationToken);
-                    lifetime = await _turnClient.RefreshAllocationAsync(lifetime);
-                    cancellationToken.ThrowIfCancellationRequested();
-                }
-                catch (OperationCanceledException e)
-                {
-                    _logger.Warning(e, $"{nameof(RefreshAllocate)}() is cancelled.");
-                    throw;
-                }
-                catch (Exception e)
-                {
-                    _logger.Error(
-                        e,
-                        $"An unexpected exception occurred during {nameof(RefreshAllocate)}(): {e}"
-                    );
-                }
-            }
-        }
-
-        private async Task RefreshPermissions(
-            CancellationToken cancellationToken)
-        {
-            TimeSpan lifetime = TurnPermissionLifetime;
-            while (!cancellationToken.IsCancellationRequested)
-            {
-                try
-                {
-                    await Task.Delay(lifetime - TimeSpan.FromMinutes(1), cancellationToken);
-                    await Task.WhenAll(Peers.Select(CreatePermission));
-                    cancellationToken.ThrowIfCancellationRequested();
-                }
-                catch (OperationCanceledException e)
-                {
-                    _logger.Warning(e, $"{nameof(RefreshPermissions)}() is cancelled.");
-                    throw;
-                }
-                catch (Exception e)
-                {
-                    var msg = "An unexpected exception occurred during " +
-                              $"{nameof(RefreshPermissions)}(): {{0}}";
-                    _logger.Error(e, msg, e);
-                }
-            }
-        }
-
         private async Task BroadcastTxAsync(
             TimeSpan broadcastTxInterval,
             CancellationToken cancellationToken)
@@ -1571,16 +1104,20 @@ namespace Libplanet.Net
             BroadcastMessage(except, message);
         }
 
-        private async Task ProcessMessageAsync(
-            Message message,
-            CancellationToken cancellationToken)
+        private void ProcessMessageHandler(object target, Message message)
         {
-            Protocol.ReceiveMessage(message);
             switch (message)
             {
                 case Ping ping:
                     {
                         _logger.Debug($"Ping received.");
+
+                        Pong pong = new Pong(BlockChain.Tip?.Index)
+                        {
+                            Identity = ping.Identity,
+                        };
+
+                        _transport.ReplyMessage(pong);
                         break;
                     }
 
@@ -1601,7 +1138,7 @@ namespace Libplanet.Net
                         {
                             Identity = getBlockHashes.Identity,
                         };
-                        (this as ISwarm).ReplyMessage(reply);
+                        _transport.ReplyMessage(reply);
                         break;
                     }
 
@@ -1625,13 +1162,17 @@ namespace Libplanet.Net
 
                 case TxIds txIds:
                     {
-                        await ProcessTxIds(txIds, cancellationToken);
+                        Task.Run(
+                            async () => await ProcessTxIds(txIds, _cancellationToken),
+                            _cancellationToken);
                         break;
                     }
 
                 case BlockHashes blockHashes:
                     {
-                        await ProcessBlockHashes(blockHashes, cancellationToken);
+                        Task.Run(
+                            async () => await ProcessBlockHashes(blockHashes, _cancellationToken),
+                            _cancellationToken);
                         break;
                     }
 
@@ -2013,7 +1554,7 @@ namespace Libplanet.Net
                 {
                     Identity = getTxs.Identity,
                 };
-                (this as ISwarm).ReplyMessage(response);
+                _transport.ReplyMessage(response);
             }
         }
 
@@ -2076,7 +1617,7 @@ namespace Libplanet.Net
             int total = hashes.Count;
             const string logMsg =
                 "Fetching a block #{Index}/{Total} ({Hash}) to include to " +
-                "a reply to {Identity}...";
+                "a reply 1to {Identity}...";
             foreach (HashDigest<SHA256> hash in hashes)
             {
                 _logger.Verbose(logMsg, i, total, hash, identityHex);
@@ -2098,7 +1639,7 @@ namespace Libplanet.Net
                         i,
                         total
                     );
-                    (this as ISwarm).ReplyMessage(response);
+                    _transport.ReplyMessage(response);
                     blocks.Clear();
                 }
 
@@ -2117,7 +1658,7 @@ namespace Libplanet.Net
                     total,
                     identityHex
                 );
-                (this as ISwarm).ReplyMessage(response);
+                _transport.ReplyMessage(response);
             }
 
             _logger.Debug("Blocks were transferred to {Identity}.", identityHex);
@@ -2236,398 +1777,8 @@ namespace Libplanet.Net
             {
                 Identity = getRecentStates.Identity,
             };
-            (this as ISwarm).ReplyMessage(reply);
-        }
 
-        private void ReceiveMessage(object sender, NetMQSocketEventArgs e)
-        {
-            try
-            {
-                NetMQMessage raw = e.Socket.ReceiveMultipartMessage();
-
-                _logger.Verbose(
-                    "A raw message [frame count: {0}] has received.",
-                    raw.FrameCount
-                );
-                Message message = Message.Parse(raw, reply: false);
-                _logger.Debug("A message has parsed: {0}, from {1}", message, message.Remote);
-                if (!(message is Ping))
-                {
-                    ValidateSender(message.Remote);
-                }
-
-                // it's still async because some method it relies are async yet.
-                Task.Run(
-                    async () =>
-                    {
-                        try
-                        {
-                            await ProcessMessageAsync(message, _cancellationToken);
-                        }
-                        catch (Exception exc)
-                        {
-                            _logger.Error(
-                                exc,
-                                "Something went wrong during message parsing: {0}",
-                                exc);
-                            throw;
-                        }
-                    },
-                    _cancellationToken);
-            }
-            catch (DifferentAppProtocolVersionException)
-            {
-                _logger.Debug("Ignore message from peer with different version.");
-            }
-            catch (InvalidMessageException ex)
-            {
-                _logger.Error(ex, $"Could not parse NetMQMessage properly; ignore: {ex}");
-            }
-            catch (Exception ex)
-            {
-                _logger.Error(
-                    ex,
-                    $"An unexpected exception occurred during ReceiveMessage(): {ex}"
-                );
-            }
-        }
-
-        private void DoBroadcast(object sender, NetMQQueueEventArgs<(Address?, Message)> e)
-        {
-            (Address? except, Message msg) = e.Queue.Dequeue();
-
-            // FIXME Should replace with PUB/SUB model.
-            try
-            {
-                var peers = Protocol.PeersToBroadcast(except).ToList();
-                _logger.Debug($"Broadcasting message [{msg}]");
-                _logger.Debug($"Peers to broadcast : {peers.Count}");
-                peers.ParallelForEachAsync(async peer =>
-                {
-                    await SendMessageAsync(peer, msg);
-                });
-
-                _logger.Debug($"[{msg}] broadcasting completed.");
-            }
-            catch (TimeoutException ex)
-            {
-                _logger.Error(ex, $"TimeoutException occurred during {nameof(DoBroadcast)}().");
-            }
-            catch (Exception ex)
-            {
-                _logger.Error(
-                    ex,
-                    $"An unexpected exception occurred during {nameof(DoBroadcast)}()."
-                );
-            }
-        }
-
-        private void DoReply(object sender, NetMQQueueEventArgs<Message> e)
-        {
-            Message msg = e.Queue.Dequeue();
-
-            // FIXME: this works, but should be fixed.
-            if (msg is Pong pong)
-            {
-                pong.TipIndex = BlockChain.Tip?.Index;
-                msg = pong;
-            }
-
-            string identityHex = ByteUtil.Hex(msg.Identity);
-            _logger.Debug("Reply {Message} to {Identity}...", msg, identityHex);
-            NetMQMessage netMQMessage = msg.ToNetMQMessage(_privateKey, AsPeer);
-
-            // FIXME The current timeout value(1 sec) is arbitrary.
-            // We should make this configurable or fix it to an unneeded structure.
-            if (_router.TrySendMultipartMessage(TimeSpan.FromSeconds(1), netMQMessage))
-            {
-                _logger.Debug("A reply sent to {Identity}: {Message}", identityHex, msg);
-            }
-            else
-            {
-                _logger.Debug("Failed to reply to {Identity}: {Message}", identityHex, msg);
-            }
-        }
-
-        private async Task ProcessRuntime(
-            CancellationToken cancellationToken = default(CancellationToken))
-        {
-            while (!cancellationToken.IsCancellationRequested)
-            {
-                _logger.Verbose("Waiting for a new request...");
-                MessageRequest req = await _requests.TakeAsync(cancellationToken);
-
-                try
-                {
-                    await ProcessRequest(req, cancellationToken);
-                }
-                catch (OperationCanceledException)
-                {
-                    _logger.Information("Cancellation requsted; shutdown runtime...");
-                    throw;
-                }
-                catch (Exception e)
-                {
-                    if (req.Retryable)
-                    {
-                        const int retryAfter = 100;
-                        _logger.Debug(
-                            $"Unexpected exception occurred during {nameof(ProcessRuntime)}(): " +
-                            "{Exception}; retry after {DelayMs} ms...",
-                            e,
-                            retryAfter
-                        );
-                        await _requests.AddAsync(req.Retry(), cancellationToken);
-                        await Task.Delay(retryAfter, cancellationToken);
-                    }
-                    else
-                    {
-                        _logger.Error("Failed to process request[{req}]; discard it.", req);
-                    }
-                }
-            }
-        }
-
-        private async Task ProcessRequest(MessageRequest req, CancellationToken cancellationToken)
-        {
-            _logger.Verbose("Request {RequestId} taken.", req.Id);
-
-            using (var dealer = new DealerSocket(ToNetMQAddress(req.Peer)))
-            {
-                // FIXME 1 min is an arbitrary value.
-                // See also https://github.com/planetarium/libplanet/pull/599 and
-                // https://github.com/planetarium/libplanet/pull/709
-                dealer.Options.Linger = TimeSpan.FromMinutes(1);
-
-                _logger.Debug(
-                    "Trying to send {Message} to {Peer}...",
-                    req.Message,
-                    req.Peer
-                );
-                var message = req.Message.ToNetMQMessage(_privateKey, AsPeer);
-                var result = new List<Message>();
-                TaskCompletionSource<IEnumerable<Message>> tcs = req.TaskCompletionSource;
-                try
-                {
-                    await dealer.SendMultipartMessageAsync(
-                        message,
-                        timeout: req.Timeout,
-                        cancellationToken: cancellationToken
-                    );
-
-                    _logger.Debug("A message {Message} sent.", req.Message);
-
-                    foreach (var i in Enumerable.Range(0, req.ExpectedResponses))
-                    {
-                        NetMQMessage raw = await dealer.ReceiveMultipartMessageAsync(
-                            timeout: req.Timeout,
-                            cancellationToken: cancellationToken
-                        );
-                        _logger.Verbose(
-                            "A raw message ({FrameCount} frames) has replied.",
-                            raw.FrameCount
-                        );
-                        Message reply = Message.Parse(raw, true);
-                        _logger.Debug(
-                            "A reply has parsed: {Reply} from {ReplyRemote}",
-                            reply,
-                            reply.Remote
-                        );
-
-                        result.Add(reply);
-                    }
-
-                    tcs.SetResult(result);
-                }
-                catch (TimeoutException te)
-                {
-                    tcs.SetException(te);
-                }
-
-                // Delaying dealer disposing to avoid ObjectDisposedException on NetMQPoller
-                await Task.Delay(100, cancellationToken);
-            }
-        }
-
-        private void CheckStarted()
-        {
-            if (!Running)
-            {
-                throw new NoSwarmContextException("Swarm hasn't started yet.");
-            }
-        }
-
-        private string ToNetMQAddress(BoundPeer peer)
-        {
-            return $"tcp://{peer.EndPoint.Host}:{peer.EndPoint.Port}";
-        }
-
-        private async Task CreatePermission(BoundPeer peer)
-        {
-            IPAddress[] ips;
-            if (peer.PublicIPAddress is null)
-            {
-                string peerHost = peer.EndPoint.Host;
-                if (IPAddress.TryParse(peerHost, out IPAddress asIp))
-                {
-                    ips = new[] { asIp };
-                }
-                else
-                {
-                    ips = await Dns.GetHostAddressesAsync(peerHost);
-                }
-            }
-            else
-            {
-                ips = new[] { peer.PublicIPAddress };
-            }
-
-            foreach (IPAddress ip in ips)
-            {
-                var ep = new IPEndPoint(ip, peer.EndPoint.Port);
-                if (IPAddress.IsLoopback(ip))
-                {
-                    // This translation is only used in test case because a
-                    // seed node exposes loopback address as public address to
-                    // other node in test case
-                    ep = await _turnClient.GetMappedAddressAsync();
-                }
-
-                // FIXME Can we really ignore IPv6 case?
-                if (ip.AddressFamily.Equals(AddressFamily.InterNetwork))
-                {
-                    await _turnClient.CreatePermissionAsync(ep);
-                }
-            }
-        }
-
-        private void ValidateSender(Peer peer)
-        {
-            if (peer.AppProtocolVersion != _appProtocolVersion)
-            {
-                DifferentProtocolVersionEventArgs args =
-                    new DifferentProtocolVersionEventArgs
-                    {
-                        ExpectedVersion = _appProtocolVersion,
-                        ActualVersion = peer.AppProtocolVersion,
-                    };
-
-                DifferentVersionPeerEncountered?.Invoke(this, args);
-
-                throw new DifferentAppProtocolVersionException(
-                    $"Peer protocol version is different.",
-                    _appProtocolVersion,
-                    peer.AppProtocolVersion);
-            }
-        }
-
-        private async Task RefreshTableAsync(
-            TimeSpan period,
-            TimeSpan maxAge,
-            CancellationToken cancellationToken)
-        {
-            while (!cancellationToken.IsCancellationRequested)
-            {
-                try
-                {
-                    await Task.Delay(period, cancellationToken);
-                    await Protocol.RefreshTableAsync(maxAge, cancellationToken);
-                    await Protocol.CheckReplacementCacheAsync(cancellationToken);
-                }
-                catch (OperationCanceledException e)
-                {
-                    _logger.Warning(e, $"{nameof(RefreshTableAsync)}() is cancelled.");
-                    throw;
-                }
-            }
-        }
-
-        private async Task RebuildConnectionAsync(
-            TimeSpan period,
-            CancellationToken cancellationToken)
-        {
-            while (!cancellationToken.IsCancellationRequested)
-            {
-                try
-                {
-                    await Protocol.RebuildConnectionAsync(cancellationToken);
-                    await Task.Delay(period, cancellationToken);
-                }
-                catch (OperationCanceledException e)
-                {
-                    _logger.Warning(e, $"{nameof(RebuildConnectionAsync)}() is cancelled.");
-                    throw;
-                }
-            }
-        }
-
-        private readonly struct MessageRequest
-        {
-            private readonly int _retried;
-
-            public MessageRequest(
-                in Guid id,
-                Message message,
-                BoundPeer peer,
-                in TimeSpan? timeout,
-                in int expectedResponses,
-                TaskCompletionSource<IEnumerable<Message>> taskCompletionSource)
-                : this(
-                      id,
-                      message,
-                      peer,
-                      timeout,
-                      expectedResponses,
-                      taskCompletionSource,
-                      0
-                    )
-            {
-            }
-
-            internal MessageRequest(
-                in Guid id,
-                Message message,
-                BoundPeer peer,
-                in TimeSpan? timeout,
-                in int expectedResponses,
-                TaskCompletionSource<IEnumerable<Message>> taskCompletionSource,
-                int retried)
-            {
-                Id = id;
-                Message = message;
-                Peer = peer;
-                Timeout = timeout;
-                ExpectedResponses = expectedResponses;
-                TaskCompletionSource = taskCompletionSource;
-                _retried = retried;
-            }
-
-            public Guid Id { get; }
-
-            public Message Message { get; }
-
-            public BoundPeer Peer { get; }
-
-            public TimeSpan? Timeout { get; }
-
-            public int ExpectedResponses { get; }
-
-            public TaskCompletionSource<IEnumerable<Message>> TaskCompletionSource { get; }
-
-            public bool Retryable => _retried < 10;
-
-            public MessageRequest Retry()
-            {
-                return new MessageRequest(
-                    Id,
-                    Message,
-                    Peer,
-                    Timeout,
-                    ExpectedResponses,
-                    TaskCompletionSource,
-                    _retried + 1
-                );
-            }
+            _transport.ReplyMessage(reply);
         }
     }
 }
