+++ Menees.Analyzers.Settings.xml
@@ -2,5 +2,5 @@
 <Menees.Analyzers.Settings>
   <MaxLineColumns>100</MaxLineColumns>
   <MaxMethodLines>200</MaxMethodLines>
-  <MaxFileLines>2900</MaxFileLines>
+  <MaxFileLines>2850</MaxFileLines>
 </Menees.Analyzers.Settings>
